requirements.txt
aiogram==3.13.1
python-dotenv
google-genai